Move custom_policy_injection test to parameters
