+++ testsuite/tests/apicast/parameters/policies/custom_policy/test_custom_policy_injection.py
@@ -1,4 +1,4 @@
-"""Test for custom policy injected through secret
+"""Test for custom policy injected through operator with secret
 """
 import pytest
 
@@ -6,13 +6,12 @@ from testsuite.capabilities import Capability
 
 
 # pylint: disable=unused-argument
-@pytest.mark.disruptive
 @pytest.mark.required_capabilities(Capability.OCP4)
-def test_apimanager_custom_policy(patch_apimanager, application):
+def test_custom_policy(patch, application):
     """
     Sends request to apicast and check that the custom policy header is there
     """
     api_client = application.api_client()
 
     response = api_client.get("/")
-    assert response.headers["X-Example-Policy-Response"] == "TEST"
+    assert response.headers.get("X-Example-Policy-Response") == "TEST"
